backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-multipart==0.0.12
pillow==11.0.0
python-dotenv==1.0.1
slowapi==0.1.9
pydantic==2.10.3
pydantic-settings==2.6.1
